remove stack-error feature flag
