+++ examples/examples/instrumented_error.rs
@@ -3,7 +3,7 @@
 #![deny(rust_2018_idioms)]
 use std::error::Error;
 use std::fmt;
-use tracing_error::{prelude::*, ErrorLayer};
+use tracing_error::{ExtractSpanTrace, boxed::{InstrumentResult, InstrumentError}, ErrorLayer};
 use tracing_subscriber::{prelude::*, registry::Registry};
 
 #[derive(Debug)]
